try-compute-engine/requirements.txt
Flask
gunicorn
tensorflow
numpy
pillow
Keras-Applications
keras-preprocessing
opencv-python
firebase-admin